refactor: separate UI and editor concerns in git diff viewer
Extract file header UI into reusable FileDiffHeader component and rename main viewer to codemirror-git-diff-viewer.tsx to make it clear this is the CodeMirror implementation. The new git-diff-viewer.tsx acts as a simple export wrapper, making it easy to swap editor implementations in the future (e.g., Monaco).
